% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.kms-package}
\alias{aws.kms-package}
\alias{aws.kms}
\title{aws.kms}
\description{
AWS Key Management Service (KMS) Client.
}
\details{
This is a client for the AWS Key Management Service (KMS).
}
\seealso{
\code{\link{create_kms_key}}, \code{\link{list_kms_keys}}, \code{\link{generate_blob}}, \code{\link{encrypt}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
