% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_kms_key.R
\name{create_kms_key}
\alias{create_kms_key}
\alias{get_kms_key}
\alias{delete_kms_key}
\title{Create/Update/Retrieve/Delete Encryption Key}
\usage{
create_kms_key(...)

get_kms_key(...)

delete_kms_key(key, ...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{kmsHTTP}}.}

\item{key}{A character string specifying a key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with \dQuote{alias/}.}
}
\description{
Create/update/retrieve/delete a KMS encryption key
}
\seealso{
\code{\link{list_kms_keys}}, \code{\link{create_kms_alias}}
}
