% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.kms-package}
\alias{aws.kms-package}
\alias{aws.kms}
\title{aws.kms}
\description{
AWS Key Management Service (KMS) Client.
}
\details{
This is a client for the AWS Key Management Service (KMS), which can be used to create and manage encryption keys used by AWS services or to setup a secure HTTP-based encryption service using \code{\link{encrypt}} and \code{\link{decrypt}}. KMS is also used natively by other AWS services.
}
\references{
\url{https://docs.aws.amazon.com/kms/latest/developerguide/overview.html}
\url{https://docs.aws.amazon.com/kms/latest/APIReference/Welcome.html}
}
\seealso{
\code{\link{create_kms_key}}, \code{\link{list_kms_keys}}, \code{\link{generate_blob}}, \code{\link{encrypt}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
